Rewrite and redesign JavaScript in the front-end
It started as a refactoring but it ended up in rewriting all the JavaScript in the front-end. This patch changes dramatically the way client code is loaded, executed and shared between different pages. It features prototypal inheritance, code organizing on modules and separate files, and no global variables. This simplifies the configuration pages of each service and it is designed to be easily extended by any new service implementation.
We are still using jQuery for now as swithcing to Closure would result in too many changes (the patch is already quite large). I tried to avoid using jQuery too much in the core components, so that switching frameworks should be easier.
Besides the JavaScript code, the patch contains changes to the user interface as well, because the functionality of polling and error reporting has to be available on every page. The polling is visualized as a decreasing timer and info/error is displayed in the top part of the page.
All the pages have been refactored, excluding the login page and TaskFarm page. These should be the subject of a future work.
Regarding the browser compatibility, theoretically it should be fully compatible with any browser. The code relies on some features from ECMAScript 5, but they are provided locally when the browser doesn't support them. Browser compatiblity is also assured by using the jQuery framework.

This is a consistent patch, so I'm attaching it as a file. For review, you can focus on the JavaScript files, especially conpaas.js, because that is the core of the new design.

git-svn-id: svn+ssh://svn.forge.objectweb.org/svnroot/contrail/trunk/conpaas@2467 ef35cc4a-2640-4bbf-98bc-65920db7d8ae