Using reCAPTCHA for the register form.
Without having a bot-detection mechanism for the register form we are vulnerable to a trivial script attack that can fill our user database with random users, leading to the front-end being unavailable. The reCAPTCHA SDK for PHP (one file - recaptchalib.php) is bundled in our code because the license is very permissive - it's the classical, one-clause BSD license.
Together with the reCAPTCHA feature, this patch includes a refactoring of the login page, including rewriting the JavaScript and the separation of the register and login AJAX endpoints, as well as a cleaner way for doing authentication and authorization.

Note for install: to allow reCAPTCHA work properly you have to generate a pair of keys on the admin page hosted by Google and you have enter those in the config.php file. For more details read config-example.php.

Reviewed By: claudiugh

Test Plan: I tried to register a new user by entering first intentionally incorrect words for reCAPTCHA, and then correct data that leads to a successful registration. I haven't tested the receiving of the Welcome email because the mail function doesn't work on cumulus.zib.de.

git-svn-id: svn+ssh://svn.forge.objectweb.org/svnroot/contrail/trunk/conpaas@3036 ef35cc4a-2640-4bbf-98bc-65920db7d8ae